Replace floats with doubles for better precision

+++ Oddity/API/Models/Common/MassInfo.cs
@@ -5,9 +5,9 @@ namespace Oddity.API.Models.Common
     public class MassInfo
     {
         [JsonProperty("kg")]
-        public float? Kilograms { get; set; }
+        public double? Kilograms { get; set; }
 
         [JsonProperty("lb")]
-        public float? Pounds { get; set; }
+        public double? Pounds { get; set; }
     }
 }
