+++ Oddity/API/Models/Common/VolumeInfo.cs
@@ -5,9 +5,9 @@ namespace Oddity.API.Models.Common
     public class VolumeInfo
     {
         [JsonProperty("cubic_meters")]
-        public float? CubicMeters { get; set; }
+        public double? CubicMeters { get; set; }
 
         [JsonProperty("cubic_feet")]
-        public float? CubicFeet { get; set; }
+        public double? CubicFeet { get; set; }
     }
 }
