+++ Oddity/API/Models/Launch/Rocket/SecondStage/Orbit/OrbitParameters.cs
@@ -9,40 +9,40 @@ namespace Oddity.API.Models.Launch.Rocket.SecondStage.Orbit
         public ReferenceSystemType? ReferenceSystem { get; set; }
 
         public OrbitRegime? Regime { get; set; }
-        public float? Longitude { get; set; }
+        public double? Longitude { get; set; }
 
         [JsonProperty("semi_major_axis_km")]
-        public float? SemiMajorAxisKilometers { get; set; }
+        public double? SemiMajorAxisKilometers { get; set; }
 
-        public float? Eccentricity { get; set; }
+        public double? Eccentricity { get; set; }
 
         [JsonProperty("periapsis_km")]
-        public float? PeriapsisKilometers { get; set; }
+        public double? PeriapsisKilometers { get; set; }
 
         [JsonProperty("apoapsis_km")]
-        public float? ApoapsisKilometers { get; set; }
+        public double? ApoapsisKilometers { get; set; }
 
         [JsonProperty("inclination_deg")]
-        public float? InclinationDegrees { get; set; }
+        public double? InclinationDegrees { get; set; }
 
         [JsonProperty("period_min")]
-        public float? PeriodMinutes { get; set; }
+        public double? PeriodMinutes { get; set; }
 
         [JsonProperty("lifespan_years")]
-        public float? LifespanYears { get; set; }
+        public double? LifespanYears { get; set; }
 
         public DateTime? Epoch { get; set; }
 
         [JsonProperty("mean_motion")]
-        public float? MeanMotion { get; set; }
+        public double? MeanMotion { get; set; }
 
         [JsonProperty("raan")]
-        public float? Raan { get; set; }
+        public double? Raan { get; set; }
 
         [JsonProperty("arg_of_pericenter")]
-        public float? ArgOfPericenter { get; set; }
+        public double? ArgOfPericenter { get; set; }
 
         [JsonProperty("mean_anomaly")]
-        public float? MeanAnomaly { get; set; }
+        public double? MeanAnomaly { get; set; }
     }
 }
